+++ SoftwareEngineering/TextVenturer/FBO.h
@@ -7,26 +7,26 @@ private:
 
     std::unordered_map<GLFBOAttachment, BaseRBO*> outputs;
 
-    size_t width;
-    size_t height;
+    UINT width;
+    UINT height;
 
     static FBO* boundFBO;
 public:
-    FBO(size_t width, size_t height);
+    FBO(UINT width, UINT height);
     virtual ~FBO();
 
-    static void bindScreen(size_t width, size_t height);
+    static void bindScreen(UINT width, UINT height);
     void bind();
 
     void enableRenderBuffer(GLFBOAttachment type, GLPixelFormat format);
-    void enableRenderBufferMS(GLFBOAttachment type, GLPixelFormat format, size_t samples);
+    void enableRenderBufferMS(GLFBOAttachment type, GLPixelFormat format, UINT samples);
 
     bool finish();
 
     void copyTo(FBO* dest, GLAttribMask mask);
     void copyToScreen(GLAttribMask mask);
 
-    void resize(size_t width, size_t height);
-    void setSamples(size_t samples);
+    void resize(UINT width, UINT height);
+    void setSamples(UINT samples);
 };
